Comment out containers
